refactor(profile): update booking query import path and adjust bookings data handling
