+++ apps/frontend/src/services/bookings/BookingQuery.ts
@@ -1,8 +1,8 @@
 "use client"
 import { useQuery } from "@tanstack/react-query"
 import { useAuth } from "@/context/AuthContext"
-import { getMyBookings } from "./BookingService"
-import { QueryKeys } from "./index"
+import { QueryKeys } from "@/services"
+import BookingService from "./BookingService"
 
 /**
  * Custom hook to fetch the current user's bookings and cache result using Tanstack Query
@@ -13,6 +13,11 @@ export function useMyBookings() {
   const { token } = useAuth()
   return useQuery({
     queryKey: [QueryKeys.MY_BOOKINGS_QUERY_KEY],
-    queryFn: () => getMyBookings(token ?? ""),
+    queryFn: async () => {
+      if (!token) {
+        throw new Error("No token provided")
+      }
+      return await BookingService.getMyBookings(token)
+    },
   })
 }
